servidor/requirement.txt
Django==1.10.4
django-rest==0.0.1
djangorestframework==3.5.3
wheel==0.24.0
